agents/python_watsonx_agent/requirements.txt
fastapi
uvicorn
python-dotenv
ibm-watsonx-ai
pydantic-settings